streamline-incident-response/requirements.txt
aws-cdk-lib==2.173.2
constructs>=10.0.0,<11.0.0
cdklabs.generative-ai-cdk-constructs==0.1.288
cdk-nag==2.34.23